% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dista.R
\name{dista}
\alias{dista}
\title{Function multiply the matrix d by an scalar a}
\usage{
dista(n, x, y, d, a)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{d}{matrix}

\item{a}{escalar}
}
\value{
matrix dij
}
\description{
This function is a toy example to call a Fortran subroutine and create a R package
}
\details{
Updated 27 December 2022
}
\author{
Carles Comas \email{carles.comas@udl.cat}
}
